-Added deleteTransaction, this method is incomplete -Added the ability to query in the audit tab based on date -Added invisible buttons for the audit tab that appear on click on a list item -Perfected interface between the entry and audit tab.  This is used when the edit button is pressed. -Modified a bunch of places where _ID was being stored as an int.  It has been changed to long to support what the DB could report for an ID.

+++ branches/MintTrackV2/MintTrack/src/com/ponyinc/minttrack/EntryActivity.java
@@ -23,7 +23,7 @@ import android.widget.TextView;
 import android.widget.Button;
 /**
  * This class represents the entry tab and all of it's pieces and interactions
- * @author Stephan Krach
+ * @author Stephan Krach & Christopher Wilkins
  */
 public class EntryActivity extends Activity {
 //	private TextView mDateDisplay;
@@ -296,7 +296,7 @@ public class EntryActivity extends Activity {
 	 * Get transaction and fill entry tab
 	 * @param trans_ID id to be used to fill tab
 	 */
-	private void setEntryTab(int trans_ID)
+	private void setEntryTab(double trans_ID)
 	{
